Refactored App.jsx content to EditorPage.jsx
This way there is room in App.jsx to create routing logic
